Move render logic into ViewComponent to allow for use of S3 docs in other parts of the application for content.
