+++ src/views/Home.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="home">
 
+    <view-component file="Home.adoc" :useS3="false"></view-component>
 
     <v-card flat class="d-flex flex-column">
       <v-card flat class="d-flex flex-row">
@@ -15,18 +16,21 @@
       <router-link to="/view/note.txt">note.txt</router-link>
     </v-card>
 
+
     <!-- <img alt="Vue logo" src="../assets/logo.png">
     <HelloWorld msg="Welcome to Your Vue.js App"/> -->
   </div>
 </template>
 
 <script>
+import ViewComponent from '../components/ViewComponent.vue'
 // @ is an alias to /src
 // import HelloWorld from '@/components/HelloWorld.vue'
 
 export default {
   name: 'Home',
   components: {
+    ViewComponent
     // HelloWorld
   },
   data() {
